Set up a flip Gravity Power up
The power up to flip gravity is set up. I just made a public method to see what the gravity is and determine what to flip it to.

Note: For whatever reason, when switching from regular gravity to inverse gravity, the player needs to jump first before it kicks in