Added Powerup interface and edited Spawner and Director
Added the IPowerup interface to allow intialzing of settings easier and condense behavior. Edited the spawner to check for if the player that interacted was last place in order to adjust the potency of powerups. Edited powerups to set a default value before being edited, if needed. Added function to Director to locate the player in first place